(CTH-01) Add basic websocket handlers and library
This commit implements basic websocket handlers via the ring-jetty9-adapter.
This requires an upgrade to Jetty 9.2 which has known issues with the
trapperkeeper-webserver-jetty9 service. For the initial spike however this
should be acceptable.

When the service starts it will accept incoming ws connections on a configurable
endpoint and port and ack any messages received from the client.

https://github.com/sunng87/ring-jetty9-adapter

This commit also changes to namespace to the modern puppetlabs.x format instead
of com.puppetlabs.x.