(cth-04) Implement connection setup steps
This commit adds the concept of a login server message. Once a
connection has been established a client has to send a login message to
the server before it can send and receive messages.

For now we're keeping the login message very simple, only a username and
client type needs to be passed. This will be extended in the future.

This commit also ups to test coverage.