CTH-5 extend get-hostname to work with ssl client certificates
Here we add a configurator builder to recompose the jetty server before we
start it when ssl is enabled.  This configurator adds Customizer
to update a map of host/port -> cert cn, so we can map from WebSocketAdapter to
CN in get-hostname.

This is all somewhat hacky and could be less so if ring.adapter.jetty9 exposed
customizer specification.  Will propose that given infinite time.