Flatten connection-map and remove endpoint-map
connection-map was { host { websocket { :type foo } } } and we had a complex
semi-inverted version of it called the endpoint-map.  This is a little weird
and doesn't really add much right now beyond book-keeping.

Flatten out connection-map so it's { WebSocket ConnectionState } and push host
(now called :client as it comes from the client certificate) down into
ConnectionState.

Removed endpoint-map and functions that did the housekeeping for it.

Replace parse-endpoints and find-sockets with a simpler
websockets-for-endpoints which simply scans forwards over the now simplified
connection-map.