Add a MeshService and hazelcast/single instance implementation
The MeshService is intended to encapsulate the behaviours of a broker which
route and deliver messages when potentially meshed.

The protocol is simple:
    (record-client-location [this client])
    (forget-client-location [this client])
    (deliver-to-client [this client message])
    (register-local-delivery [this fn])

We have two implementations, single and hazelcast.

`single` as not actually meshed it just connects the 'deliver this
message' to the 'you have a message to deliver' callback and gets
out the way.  It doesn't care for record-location or forget-location.

The hazelcast implementation

    Assigns a UUID to the broker
    creates a MultiMap { broker-id [ client-id client-id ] } (the location map)
    subscribes to a queue named for the broker-id

When dispatching a message it searches the location map for a broker
that endpoint is connected to (any, future improvement would be to
identify it's connected locally and bypass hazelcast), and then
puts it on the queue for that broker.