(CTH-55) Add inventory message to cthun server
This commit adds the concept of an inventory server-message.

A controller will send a message with a :query field using the
http://puppetlabs.com/inventoryschema data_schema. The query will
then be processed. Right now that is limited to wildcard expansion
but might be puppetdb queries in a future implementation.

The server will then respond with a message with an :endpoints field
which will contain a vector of endpoints returned by wildcard expansion.