(CTH-74) Add JMX metrics
This commit uses metrics-clojure to expose basic metrics from the cthun
server internals.

First round of values are:
- total active connections
- number of messages received
- number of messages sent
- the rate of messages sent
- the rate of messages received

We also expose 3 timings that we can use to determine the amount of time
spent in the onConnect, onText and onClose callbacks.

Note that these metrics aren't free and we should keep a close eye on
how they impact furture performance.

This commit also adds an endpoint on port 3000 which will respond with
some of the metrics as a json document. This will become fully baked in
a future commit.