maint - comment out broken, mistaken, tests
The tests in websockets-test have always tested slightly too much, as they test
the behaviour of the things they call on at too far a remove.

For now leave the tests as comments, they should be removed at the earliest
moment once we're sure the methods we're calling have testing in the correct
places.