(maint) move explode-endpoint into validation namespace
Here we move the explode-endpoint utility function over into the
validation namespace, so the next step we take (refactoring
check-certname in terms of it) doesn't cause a recursive require
loop.

It may be worth moving this over into it's own namespace at some
later point, if we have other endpoint-related operations.