CTH-120 Add message abstraction
Here we add additional functions to the message namespace to encapsulate
decoding from a network buffer, and encoding for a network buffer.

It's a very thin abstraction over the existing message frame, adding only a
private :_destination field which we'll use when expanding a message into a
sequence for multiple delivery.