CTH-141 do not allow the same endpoint to be connected multiple times
Previously we had no guard for the same actor (identified by certificate+role)
being logged in from two websockets, instead this would cause odd behaviours
interfering with the messages delivered to one or both of the duplcate actors.

Here we check if we have a websocket-of-endpoint as part of processing login,
and if we detect a duplicate we reject the new connection by closing it.