CTH-147 replace message format with CTH-100 binary format
Here we change our network format from a simple json document to a chunked
binary format described in CTH-100.

This exposes a few places where the message abstraction added in CTH-120 was
not quite comprehensive enough, so we add additional helpers to the message
namespace and refactor to use it.