CTH-176 add support for destination reports
Here we add a new mechanism to message processing.  If the sender
of a message has specified the optional 'destination_report' flag
as being true in the message envelope then the server generates a
new message of the 'http://puppetlabs.com/destination_report' schema
disclosing where the message will be sent.