CTH-187 use the extracted cthun-message dist
Here we delete our original copy of src/puppetlabs/cthun/message.clj and make
use of the version available from internal nexus server.

As part of extracting the message abstraction, the validation for Envelopes was
moved from the validation to the message namespace.  It is expected this will
contnue eventually removing the validation namespace and moving validation
closer to the thing validated.