(MAINT) Add user repl namespace
This commit does the following:

* Small tweaks to lifecycle management of Jetty server, such that
  it does not block on startup (and thus the TK `start` lifecycle
  function returns control), and adds `stop` functionality to
  shut Jetty down cleanly.
* Adds 'user.clj' namespace, sets it as the default for the REPL.
  This allows you to start the app directly from the REPL (via `(go)`).
  You can also call `(stop)` to stop the app, or `(reset)` to reload
  the Clojure code and restart the app without restarting the JVM.

I haven't tested the REPL lifecycle stuff with a real live client;
have only validated that I can successfully reload the app multiple
times in the same REPL session without errors.  It's possible
that some of the global state might not quite work properly with
this pattern.