CTH-249 Allow the later session to connect
Previously when a connection attempted to associate with a client
URI that was already active within the network a second attempt to
associate with that URI was denied, as there could only be one.
This behaviour however isn't useful in the case where a client loses
its connection in an unclean manner and needs to re-associate before
the server notices that the previous client connection has terminated.

Here we change the logic to allow the later session association of
the same client URI to disconnect the old connection and usurp the
identity.