(RE-5212) Switch project build-type to PE
Previously we were using the implied FOSS type of ezbake. Thus we were
buildling for many platforms, but not the ones we need for PE
integration.

This commit flips the build-type to PE, which should allow packages to
come out for all platforms needed for our next feature release of Puppet
Enterprise.