(maint) Remove the executor service
We suspect executor service of being a potential point of failure,
and we know for sure that it isn't covered by a persistent transaction,
so here we refactor to use an activemq managed queue for delivery.

Previously the message flow was:

  Websocket -> Queue[accept] -> (expand targets) -> ExecutorService -> Websocket(s)

Now the message flow is:

  Websocket -> Queue[accept] -> (expand targets) -> Queue[delivery] -> Websocket

We also refactor a little to stop referring to queues as topics,
and improve log messages.