CTH-303 Refactor on-message! to present a state machine
Here we refactor the flow of messages from `on-message!` outwards to
make the states of a connection and the message flow at those states
more obvious and uniform.

We do this by making a more explicit state machine for Connection
objects.  Transition functions all share the signature ([Broker
Capsule Connection] :- Connection) where the return value is the
next state of the Connection.

At the time of writing this may feel like overkill as there are
only two valid states (:open and :associated), but we expect that
version negotiation will fit in is a transition to a state between
the current :open and :associated states.

Supporting changes:
- ConnectionState schema extracted from the existing (:state Connection)
  schema.
- Connection now tracks websocket.
- `new-socket` is renamed to `make-connection`, and now takes a websocket
  as a parameter.
- Schema added to return value of add-connection!, which is now returns a
  Connection.
- Session association is simplified, `association-response` is now
  `reason-to-deny-association` and returns `(s/maybe s/Str)` and in the regular
  case `process-associate-message` just returns the 'associated'
  Connection (on success).
- `process-message` melts into the state machine.
- Broker now contains a `transitions` map which is a mapping from
  ConnectionState to Transition Function.