CTH-328 Track message movement through the broker in a Capsule
Here we introduce a new abstraction
puppetlabs.cthun.broker.capsule/Capsule.  A Capsule is a container
for a message as the broker processes it, meaning we no longer need
to store broker state in the Message object.

Yes Capsule is probably a bad name, the obvious name - Envelope -
also refers to the first chunk of a Message so there was a clash
there.  Other names considered were:

    MessageState BrokerMessage BrokerEnvelope Frame MessageWrapper
    Container MessageContainer

Unit tests are now really showing their ugliness and need a rework
but they all pass.