(CTH-134) Message hops should include the server identity
In the past the server URI was hardcoded to be pcp:///server, the URI
used by clients when addressing the server.

Here we update the broker-uri fn to return the correct server URI based
on the cn in it's ssl-certificate. The ssl-cert is determined by looking
up the server value in the :websocket route, or directly from the server
if there is only one configured.