(maint) Allow for common-name of a connection to be nil
In the case where a connection did not provide an acceptable peer certificate,
common-name will be nil.  There were places where we expected common-name would
always be a string, so update the schemas.

As part of this we refactor connection/make-connection a little to
extract the common name as part of wrapping the websocket, so we
can drop the use of get-cn at log time, this is a slight performance
win.