PCP-124 account for potential race in test
We have a slightly unreliable test here, sometimes in a race netty (the
websockets client) will see that the connection has been shutdown before it
sees the reason for shutdown (code 4003 - you have no certificate).

This can be reduced by introducing this unwanted unwanted voodoo sleep in
the brokers on-connect handler, which seems to give netty enough
time to see it established a websockets session which will then be shutdown:

    diff --git a/src/puppetlabs/pcp/broker/core.clj b/src/puppetlabs/pcp/broker/core.clj
    index c8426eb..a255c68 100644
    --- a/src/puppetlabs/pcp/broker/core.clj
    +++ b/src/puppetlabs/pcp/broker/core.clj
    @@ -374,6 +374,8 @@
                    (sl/maplog :debug (assoc (connection/summarize connection)
                                            :type :connection-no-peer-certificate)
                                "No client certificate, closing {remoteaddress}")
    +
    +                 (Thread/sleep 250)
                    (websockets-client/close! ws 4003 "No client certificate"))
                    (do
                    (websockets-client/idle-timeout! ws idle-timeout)

Intead of taking this approach, we just accept that sometimes we'll get a 1006
close-code in this test, and leave breadcrumbs to the ticket where
other notes have been gathered.