PCP-146 Reorganise tests for independent selection
lein-cloverage doesn't support test selectors and always loads all
tests from the :test-paths of a project, ahead of any user-specifiable
filtering, so here, in order to add the ability to get coverage
reports on the two test subsets (unit and integration), we rearrange
the tests and add two additional leiningen profiles.

To run coverage for all our tests:

    lein coverage

Just the unit tests:

    lein with-profile +unit coverage

Just the integration tests:

    lein with-profile +integration coverage

To clarify, these integration tests integrate an entire trapperkeeper
service stack with real websockets client code.  They are still
self-contained to the repository and so may differ with your local
definition of what an integration test is.