(maint) Fix timing sensitive test
The previous expression of this test made some unsound assumptions about the
passage of time.  Moving between various runs sometimes the 9 seconds
yielding an exactly 4 seconds delay was incorrect and instead returned the
fraction 9/2, or 4.5.  Here we check that the computed value fits into the
non-descending sequence of expected values, so we're less tied to actual
clocks, without mocking out all the time logic.