PCP-193 Set up the trapperkeeper-status handler
Here we integrate trapperkeeper-status as a service dependency, and implement
broker.core/status to expose a similar set of performance metrics that the
current 'metrics' webapp does.

The main difference here is that we follow the pattern of only replying with
extremely course information at level 'critical', adding in broker statistics
at 'info', and JVM memory/thread statistics at 'debug'.