PCP-195 Supply "in-reply-to" header when responding to a message
The "in-reply-to" envelope property was initially added to the
inventory_response codepath in commit 92b66d1c.  Here we change the
behaviour of the broker to supply this property whenever we are
clearly responding to a message, allowing a client to correlate a
response with its original request.

We also note that error_message is overloaded and is used for 3
distinct error paths currently - malformed message, unauthorised
message, and internal processing error. In the latter two cases we
are able to supply an in-reply-to header.