PCP-250 Add multiple protocol version support
We dub the developing version of the protocol "vNext" for now, as
it is a developing set of changes to the already established PCP
v1.0, and give it the websockets route :vNext.

Due to the closeness of v1.0 and vNext (vNext currently just adds
the optional in-reply-to header to all reply types), the brokers
internal representation of messages is vNext, which we then transform
when sending/receiving message via a Codec which is a property of
the connection derived from the uri the connection was established
on.