(maint) Remove naming conflicts with PuppetDB
This commit renames the namespaces copied from PuppetDB so that they are
unique to this project. This avoids conflicts with PuppetDB that arise
from having two (now different!) namespaces with the same name.

Similarly, changed the name of the embedded broker from "localhost" to
"pcp" to allow for multiple ActiveMQ instances to be started in the same
process.

These changes are meant to enable pcp-broker to be embedded in the same
jar/process as PuppetDB. Eventually this code probably ought to be
migrated to shared libraries or a dedicated trapperkeeper service (for
ActiveMQ), but for now this avoids conflicts.