(PCP-448) Use ConcurrentHashMap for tracking connections
Many Jetty threads can result in modifying the connections or uri map.
When many updates occur concurrently, we can see lots of failed atom
updates resulting in much slower processing.

Switch to using a ConcurrentHashMap. It appears to scale better, and is
at least as fast for smaller numbers of clients.