(PCP-467) On demand schema validations
This commit introduces means of turning on schema validations on demand
as a preparation for removal of the `^:always-validate` attribute from
function definitions in this project (which effectively always forced
the validations).

There is a new lein alias - `tkv` - added which is identical to the
`tk` alias except it enables the schema validations before it runs
trapperkeeper.
The tests are updated to enable the validations by using the
`schema.test/validate-schemas` fixture.