(PCP-487) Increase timeouts in tests seriously
This commit increases expiry timeouts on PCP messages exchanged during
tests by a factor of 60. This is to avoid occasional failures in CI
we keep seeing.
In contrast to previous attempts to do the same this commit doesn't
change the actual timeouts but rather redefines the function which
applies the timeout values to the PCP messages:
`puppetlabs.pcp.message/set-expiry`.
This has the desired effect of raising the timeouts for both the test
code and the tested code at the same time (previous attempts were only
raising the timeouts in the test code).
Additionally, timeout is also raised in the
 `puppetlabs.pcp.testutils.client/recv!`
function which, too, was causing test failures by not waiting long
enough for messages to be received.
Finally the `it-closes-connections-when-not-running-test` test is
redesigned such that it avoids another instance of not waiting long
enough for the broker to respond.