[WIP] (PCP-489) Refactor and improve onMessage handling
Still WORK IN PROGRESS as:
 - I must update the integration tests;
 - I may add new tests to core_test;
 - there are a number of TODOs to be resolved;
 - I will try to factor out PCP-385 changes and split the commit in
   different pieces.

Checking the validity of incoming messages only once, meaning:
 - ignore messages other than associate_request before Association is
   complete;
 - authentication;
 - authorization;
 - TTL check.

Removing the transitions map to keep track of the state of the
association during the message double dispatch mechanism in favour of
making explicit the processing workflow.

Avoiding to update the 'connections' map for each processing message. In
practice, the onMessage handle will update the map only if the
Connection instance being processed is actually updated; that happens
only in case of a successful associate_request.

In case of invalid (as above) associate_request, an unsuccessful
associate_response is sent back to the requester and the WebSocket
connection is closed. That's a behaviour change, as:
 - in case of authentication failure, the broker used to send a PCP
   error and close the WebSocket;
 - in case of authorization failure, the broker used to only close the
   WebSocket.