(PCP-526) Don't test error when wrapping a bad Message
After disabling schema validation, capsule/wrap does not enforce that
the Message argument is a valid Message, so we can't test that anymore.

Note that with PCP-525 we will ensure that only valid PCP messages are
accepted by pcp-broker; we can safely drop the Message validation when
creating a Capsule.