(PCP-489) Improve onMessage handling
Checking the validity of incoming messages only once, meaning:
 - ignore messages other than associate_request before Association is
   complete;
 - authentication;
 - authorization;
 - TTL check.

Removing the transitions map that keept track of the state of the
association during the message double dispatch mechanism in favour of
making explicit the processing workflow.

Avoiding to update the 'connections' map at each processed message. In
practice, the onMessage handler will now update the map only if the
Connection instance being processed is actually updated; that happens
only in case of a successful associate_request.

Ensure incoming messages are valid Message instances by explicitly
validating against the Message schema; this is necessary as we don't
have schema validation enabled on function calls.

Removing schema validation calls for data content created by the broker.
Factoring out functions that create data content objects in order to
enable schema validation only when globally set (with appropriate lein
profile).

In case of invalid (as above) associate_request, an unsuccessful
associate_response is sent back to the requester and the WebSocket
connection is closed. That's a behaviour change, as:
 - in case of authentication failure, the broker used to send a PCP
   error and close the WebSocket;
 - in case of authorization failure, the broker used to only close the
   WebSocket.

Improving documentation.

Updating tests and creating new ones.

Using the async 'put!' call in testutils/client.clj as the previous
'>!!' blocking function was creating a deadlock in those cases where the
test client was expecting both PCP and WebSocket responses from the
broker under test.