(maint) Use Connection's :remote-address in make-ring-request
Using the :remote-address instead of invoking Jetty methods on the
WebSocket instance. Note that the remote-address string provided by the
Connection object will give only the IP address and port; there will be
no hostname entry and the '/' separator (trapperkeeper's authorization
service seems to be happy with that).

Change for PCP-385 review: removing the pcp_access entry for processing
errors.