(PCP-597) Get broker name from webserver context
Previously the broker name was read from the certificate specified in the ssl
configuration of the webserver hosting the broker. This is not possible any
more as:
- ssl may not be configured for the webserver at all
- the ssl configuration actually used by the webserver may differ from that
  specified in its configuration (it can be overridden)

So newly we try to obtain the broker name from the ssl-context actually used
by the webserver hosting the broker and if that fails we fall back to using
the hostname of the host running the broker.