(ORCH-1551) Validate message type & targets to avoid unexpected characters
Previously, any value was allowed as a message_type, including a string
containing embedded null characters. This could be used to fool the
tk-auth check and pass the message along to the client. That requires
the client to also properly handle embedded null characters or
potentially misinterpret the message_type, fully bypassing the auth
check.

We now restrict the list of characters that can be used in a
message_type to only a common set of ASCII characters that covers the
existing URI-formatted message types. While at it we add a similar check
for the message targets too.