(maint) Dynamically compute remote-address and common-name
Remote address and common name are computed from an active connection.
In preparation for outgoing pcp-client connections, return these to
being computed dynamically. The reason they were originally changed was
to allow printing the remote address when a client disconnects; in
retrospect this doesn't seem particularly useful as it can still be
retrieved from the connection message.