(PCP-600) Fix race sending inventory update before response
Normally a client requesting updates is registered as a subscriber, and
an inventory response sent. It happens in this order to ensure no
updates are missed; the reverse order would allow for an inventory
response to be sent then a change in inventory to happen before the
subscription is registered.

However, the current order allows for an inventory update to be sent by
the subscription thread between registering as a subscriber and sending
the inventory response. Register the subscription, but prevent sending
any updates until the inventory response is sent.