(PCP-701) Fix state to report starting and stopping
In a previous commit, status was changed to report error if all
controllers were disconnected. However, this is also normal behavior
when starting or stopping the service, which made status less
illuminating. Restore reporting starting and stopping status, and only
report error if the broker is running and all controllers are
disconnected.