(PE-19981) Bump to clj-parent 0.6.0
Bump to clj-parent 0.6.0 which includes trapperkeeper-metrics 1.0.0.

No code changes are required for this update.

The `:enabled` setting in the test metrics config was removed as this setting
had been deprecated and acted as a no-op for a long time; tk-metric 1.0.0
removed it entirely.

The metrics.conf included in packaging was updated to match the new config for
tk-metrics 1.0. As with the previous metrics.conf, it enables the JMX metrics
reporter. It also includes commented out example config for configuring the
Graphite reporter. The updates to this config file are not expected to cause
any problems because 1) it was never configurable using the PE module, and 2)
there is no documentation for how to configure any settings in it.