(PE-19981) Fix schema error caused by metrics.conf
HOCON interprets a `.` in a key as two keys - so `puppetlabs.orchestrator` was
being interpreted as `puppetlabs { orchestrator { ... }}` which violated the
metrics service config schema. This commit fixes the problem by wrapping the
domain in quotes, so that the `.` is interpreted literally. HOCON still treats
a string key as a keyword, so the metrics schema does not have a problem with
it.