(maint) Update to structured logging 0.2.0
Structured logging 0.2.0 introduced a breaking change. The last argument
of maplog is now expected to be a function that accepts the structured
log data as an argument. Formatting is no longer handled by maplog.
Update use of maplog to use functions and handle formatting.

Pickup the structured-logging version from clj-parent 0.8.1.