(PCP-737) Handle events on stale connection gracefully
This commit makes sure errors or messages received on stale websocket
connections are handled gracefully.
A stale connection in this context means a connection which is not present
in the broker's inventory any longer.
Previously it was not the case which could lead to validation errors
when there was an attempt to log the connection details for connection
which was not found in the broker's inventory.