(PCP-737) Handle events on stale controller connections too
A recent commit made sure events on stale connection are handled gracefully.
(A stale connection meaning a connection which is not present in the broker's
inventory any longer.) But it failed to address similar problem on controller
connections which are kept in a separate inventory / registry.
So this commit makes sure even events on stale controller connections are
handled gracefully.