(PCP-862) Remove close! call
This removes the `close!` call before disconnecting a superseded connection. It was originally left to preserve the code and reason in logging, but appeared to leave unused sockets open, consuming resources even when disconnect was called. Using just `disconnect` frees the resources, and we determined there was already sufficient logging of pxp-agents being superseded to lose the 'Superseded' message.