(maint) Lower default idle-timeout from 15 minutes to 6
Previously, we waited 15 minutes for a connection to idle out. However,
the pxp-agent is configured to send a ping event every two minutes and
to give up after six minutes of failed ping attempts. This lowers the
idle-timeout to six minutes to better match pxp-agent configuration. See
PCP-729 for discussion of pxp-agent changes.