(PCP-900) Close connections after CRL update
tk-jetty9 will reload the SSLContextFactory for pcp-broker when changes
to the CRL occur. In doing so new connections will use the CRL updates.
However pcp-broker, by design, keeps connections open for long durations.
This can result in many different SSLContexts in the heap with different
versions of the CRL. Coupled with large CRLs this can take up a
significant portion of the heap (10-20%).

This brings in tk-filesystem-watcher as a dependency and, like
tk-jetty9, watches the CRL for updates. When it has changed the client
are disconnected fromt the server. This is under the assumption that the
clients will reconnect and the server will be able to use the new CRL.

Some caveats:

 - The SSLContextFactory is not guaranteed to have reloaded by the time
   the first agent reconnects. This might cause 2 CRL versions in memory,
   but should limit the number to 2.  Unfortunately, even though
   SSLContextFactory implements a Lifecycle listener pattern, reloads do
   not seem to trigger them (the code looks like only start and stop are
   implemented).

 - This could cause significant churn in large deployments.