(maint) Update SSL test fixtures
Newer versions of Java will not allow CA certs that don't have the
"CA:TRUE" extension. This means all the certs must be recreated, and it
is just as easy to then make new private and public keys as well. This
commit also provides a script to recreate the SSL infrastucture on
demand and a readme on how to do so.